+++ ffi/diplomat/tests/missing_apis.txt
@@ -31,9 +31,8 @@ icu::calendar::DateTime::try_new_persian_datetime#FnInStruct
 icu::calendar::DateTime::try_new_roc_datetime#FnInStruct
 icu::calendar::DateTime::try_new_ummalqura_datetime#FnInStruct
 icu::calendar::chinese::Chinese#Struct
-icu::calendar::chinese::Chinese::chinese_new_year_on_or_before_iso#FnInStruct
-icu::calendar::chinese::Chinese::major_solar_term_from_iso#FnInStruct
-icu::calendar::chinese::Chinese::minor_solar_term_from_iso#FnInStruct
+icu::calendar::chinese::Chinese::EPOCH#AssociatedConstantInStruct
+icu::calendar::chinese::Chinese::location#FnInStruct
 icu::calendar::chinese::ChineseDateInner#Struct
 icu::calendar::hebrew::Hebrew#Struct
 icu::calendar::hebrew::Hebrew::new#FnInStruct
